Corrección de screens Events y MyEvents
