+++ src/screens/Events/styles.js
@@ -1,3 +1,8 @@
 import { StyleSheet } from 'react-native';
 
-export default styles = StyleSheet.create({});
+export default styles = StyleSheet.create({
+  container: {
+    marginTop: 50,
+    alignItems: 'center',
+  },
+});
